django_app/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.6
django-admin-interface==0.28.6
django-admin-tools==0.9.3
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
zope.interface==6.4.post2
